feat: init swc
