+++ package.json
@@ -23,6 +23,7 @@
     "@babel/runtime-corejs3": "^7.12.5",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
     "@svgr/webpack": "^5.5.0",
+    "@swc/core": "^1.2.197",
     "babel-loader": "^8.2.1",
     "copy-webpack-plugin": "^6.3.1",
     "core-js": "^3.7.0",
@@ -47,6 +48,7 @@
     "semver": "^7.3.2",
     "source-map-loader": "^3.0.0",
     "style-loader": "^2.0.0",
+    "swc-loader": "^0.2.1",
     "ts-loader": "^9.2.4",
     "typescript": "~4.1.3",
     "walk": "^2.3.14",
