+++ preset-configs/default.json
@@ -66,7 +66,8 @@
     "transformDeps": false,
     "highQualitySourceMap": false,
     "filesystemCache": true,
-    "errorOverlay": true
+    "errorOverlay": true,
+    "swc": true
   },
   "devProxy": {},
   "deploy": {
